feat: 컴포넌트 분리 및 컴포넌트 배치 변경
