+++ client/src/features/EventManage/components/GuestManageSection.tsx
@@ -2,21 +2,14 @@ import { useState } from 'react';
 
 import { css } from '@emotion/react';
 
-import { Button } from '@/shared/components/Button';
-import { Card } from '@/shared/components/Card';
 import { Flex } from '@/shared/components/Flex';
-import { Icon } from '@/shared/components/Icon';
-import { Input } from '@/shared/components/Input';
-import { Text } from '@/shared/components/Text';
 import { useModal } from '@/shared/hooks/useModal';
 
-import { GuestList } from './GuestList';
-import { GuestModal } from './GuestModal';
+import type { Guest } from '../types/Guest';
 
-export type Guest = {
-  name: string;
-  status: string;
-};
+import { AlarmSection } from './AlarmSection';
+import { GuestModal } from './GuestModal';
+import { GuestViewSection } from './GuestViewSection';
 
 type GuestManageSectionProps = {
   completedGuests: Guest[];
@@ -26,7 +19,6 @@ type GuestManageSectionProps = {
 export const GuestManageSection = ({ completedGuests, pendingGuests }: GuestManageSectionProps) => {
   const { isOpen, open, close } = useModal();
   const [selectedGuest, setSelectedGuest] = useState<Guest | null>(null);
-  const [alarmMessage, setAlarmMessage] = useState('');
 
   const handleGuestClick = (guest: Guest) => {
     setSelectedGuest(guest);
@@ -38,93 +30,34 @@ export const GuestManageSection = ({ completedGuests, pendingGuests }: GuestMana
     setSelectedGuest(null);
   };
 
-  const handleSendAlarm = () => {
-    if (alarmMessage.trim()) {
-      // TODO: 알람 전송 로직 구현
-      console.log('알람 전송:', alarmMessage);
-      setAlarmMessage('');
-    }
-  };
-
   return (
-    <Flex as="section" dir="column" gap="24px" width="100%">
-      <Card>
-        <Flex dir="column" gap="20px">
-          <Flex alignItems="center" gap="8px">
-            <Icon name="users" size={14} color="#4A5565" />
-            <Text type="Body" weight="regular" color="#4A5565">
-              게스트 조회
-            </Text>
-          </Flex>
+    <Flex
+      as="section"
+      dir="column"
+      gap="24px"
+      width="100%"
+      margin="10px"
+      css={css`
+        max-width: 800px;
+        margin: 0 auto;
+        padding: 0 16px;
 
-          <GuestList
-            title={`신청 완료 (${completedGuests.length}명)`}
-            titleColor="#00A63E"
-            guests={completedGuests}
-            variant="completed"
-            onGuestClick={handleGuestClick}
-          />
+        @media (max-width: 768px) {
+          padding: 0 20px;
+        }
 
-          <GuestList
-            title={`미신청 (${pendingGuests.length}명)`}
-            titleColor="#4A5565"
-            guests={pendingGuests}
-            variant="pending"
-            onGuestClick={handleGuestClick}
-          />
-        </Flex>
-      </Card>
+        @media (max-width: 480px) {
+          padding: 0 16px;
+        }
+      `}
+    >
+      <AlarmSection pendingGuestsCount={pendingGuests.length} />
 
-      <Card>
-        <Flex dir="column" gap="16px">
-          <Flex alignItems="flex-end">
-            <Icon name="alarm" size={24} color="#F54900" />
-            <Text type="Body" weight="medium" color="#F54900">
-              미신청자 알람
-            </Text>
-          </Flex>
-
-          <Input
-            id="alarm-message"
-            label=""
-            placeholder="알람 메시지를 입력하세요..."
-            value={alarmMessage}
-            onChange={(e) => setAlarmMessage(e.target.value)}
-            css={css`
-              & input {
-                background-color: #f3f3f5;
-                border: none;
-                border-radius: 6.75px;
-                padding: 8px 12px;
-                font-size: 12.3px;
-                color: #717182;
-                &::placeholder {
-                  color: #717182;
-                }
-              }
-            `}
-          />
-
-          <Button
-            width="100%"
-            size="sm"
-            color="#F54900"
-            disabled={!alarmMessage.trim()}
-            onClick={handleSendAlarm}
-            css={css`
-              border-radius: 6.75px;
-              opacity: ${alarmMessage.trim() ? 1 : 0.5};
-              cursor: ${alarmMessage.trim() ? 'pointer' : 'not-allowed'};
-            `}
-          >
-            보내기
-          </Button>
-
-          <Text type="caption" weight="regular" color="#6A7282">
-            {`${pendingGuests.length}명의 미신청자에게 알람이 전송됩니다.`}
-          </Text>
-        </Flex>
-      </Card>
+      <GuestViewSection
+        completedGuests={completedGuests}
+        pendingGuests={pendingGuests}
+        onGuestClick={handleGuestClick}
+      />
 
       <GuestModal isOpen={isOpen} onClose={handleCloseModal} guest={selectedGuest} />
     </Flex>
